Big Changes to Read_My_qPCR
Completely changed the way headers are handled. Now, headers are split by " - " delimiter (and it needs to stay this way).
Also, what goes into the plots is inputted by user at the beginning of code running.

Also, added a defined function for making subplots and updated # PLOT 1 and # PLOT 2 to use this.

#PLOT 3 and later need to be reworked. They are currently defunct.